chore: Breakout enums from type file
