+++ packages/react-paypal-js/src/v6/components/PayPalProvider.ssr.test.tsx
@@ -8,7 +8,7 @@ import { loadCoreSdkScript } from "@paypal/paypal-js/sdk-v6";
 
 import { PayPalProvider } from "./PayPalProvider";
 import { usePayPal } from "../hooks/usePayPal";
-import { INSTANCE_LOADING_STATE } from "../types/PayPalProviderTypes";
+import { INSTANCE_LOADING_STATE } from "../types/PayPalProviderEnums";
 import { isServer } from "../utils";
 import { TEST_CLIENT_TOKEN, expectInitialState } from "./providerTestUtils";
