+++ packages/react-paypal-js/src/v6/hooks/usePayPal.ts
@@ -2,7 +2,7 @@ import { useContext } from "react";
 
 import { PayPalContext } from "../context/PayPalProviderContext";
 
-import type { PayPalContextState } from "../types/PayPalProviderTypes";
+import type { PayPalContextState } from "../types/PayPalProviderTypes.d.ts";
 
 export function usePayPal(): PayPalContextState {
     const context = useContext(PayPalContext);
