+++ packages/react-paypal-js/src/v6/types/PayPalProviderTypes.d.ts
@@ -1,19 +1,8 @@
 import type { Components, SdkInstance, EligiblePaymentMethodsOutput } from ".";
-
-export enum INSTANCE_LOADING_STATE {
-    INITIAL = "initial",
-    PENDING = "pending",
-    RESOLVED = "resolved",
-    REJECTED = "rejected",
-}
-
-export enum INSTANCE_DISPATCH_ACTION {
-    SET_LOADING_STATUS = "setLoadingStatus",
-    SET_INSTANCE = "setInstance",
-    SET_ELIGIBILITY = "setEligibility",
-    SET_ERROR = "setError",
-    RESET_STATE = "resetState",
-}
+import type {
+    INSTANCE_LOADING_STATE,
+    INSTANCE_DISPATCH_ACTION,
+} from "./PayPalProviderEnums";
 
 export interface PayPalState {
     sdkInstance: SdkInstance<readonly [Components, ...Components[]]> | null;
