+++ packages/react-paypal-js/src/v6/types/index.ts
@@ -1,2 +1,3 @@
 export type * from "@paypal/paypal-js/sdk-v6";
-export * from "./PayPalProviderTypes";
+export type * from "./PayPalProviderTypes";
+export * from "./PayPalProviderEnums";
